Allow price update in parallel with other javascript actions
The javascript code to provide an updated price took control of any equivalent element action rather than sharing the action(s). This update (not compatible with IE8 and below) provides that additional capability so the load sequence of this script does not need to be coordinated with other script provided the other script doesn't affect the same data.