Corrections to DPU Class
Modified to support PHP 7.2.
Modified output to account for decimals associated
 with the chosen currency.
Typecast $value as an integer when requesting the
products_options_values_name from the database using
a sql statement instead of a ZC built-in function.
Was getting repeatedly logged errors/warnings until
the product was selected and in the cart. :).