Add support to use ZCs ajax loader
This somewhat replaces the dpu_ajax.php file with the ZC ajax loader, although
really the dpu_ajax file just is reused under a different area.  Basically
this adds a level of security to the plugin and should improve update
consistency.  The recent version of the ajax.php file as provided in
ZC 1.5.5e is included with this commit.