Implement zcJS and other minor code improvements.
Added JSON response information when retrieving results.
Incorporated the use of Zen Cart's provided zcJS when it is available.
Made the DPU class part of the overall system to better support additional observers as needed.
Maintained the "default" XML response if JSON is not requested to align with potential historical alterations.
Added the restoration of the price display if the transaction fails.
Added jscript function to perform data/screen update/swap.