Rewrote the data collection about a product to improve ability to compare to the active cart.
If using the class function of insertProducts,
the resulting prid is an integer only and does not
include the possibility of addressing attributes
based on the change(s) made here.  This is equivalent
to a standard product being listed in the product listing
screen being added to the cart where attributes are
"typically" not selectable.
Modified quantity reporting to provide quantity already
in cart for selected attributes if the quantity is
greater than zero.