Restore temporary cart attributes.
Removes from the temporary cart the attributes that
were added to support identifying the remaining
minimum charge based on other attributes that
were not selected but likely would be expected to
be added.

Perhaps there could be a case developed of supporting
say addition of a product that has radio buttons
but none of them are selected when adding to the cart;
however, this is not default behaviour and as such
DPU is not designed to specifically handle that case
but instead to handle the default case of say a
radio attribute option name being selected to
support determination of the lowest remaining price
based on the remaining possible selections.