Include page information with ajax request
Has been identified that selecting an attribute
from the page causes ZC to process the request as
a new page request and therefore programs that
track the status (including who is online) report
the ajax call as the current page for the customer
This adds the general page information (though not
with zen_href_link) and support understanding what
the customer is looking at at the point of the call.

Also a minor update in javascript declaration.

Updated the product_music_info jscript file.