Support priced-by-attribute-product and less than total selection of product.
Adds the ability to determine if when selecting attributes for a product
that is priced-by-attributes whether the needed selections have been
made or if there remain options to be selected.

It also identifies the minimum price to be paid for the same product
that is priced-by-attributes based on the lowest remaining price to be
selected shows on the product information page.

This does not account for product that are not priced by attribute where
specifically a choice must be made in order to add the product to the
cart (such as a drop down with all allowable options adding a price
to the cart as compared to a checkbox that is unselected).