Fixes #5 (at least the last identified part of it)
Other portions of the price display have been corrected; however, this
at least allowed display of the Golf Clubs (products_id=132) in the
demo product catalog to display a DPU base price based on not selecting
any option, then increasing the price for each selection added.

In ZC versions older than 1.5.6 there is the possibility that this will not
display properly if the sum of the options_values_price for a product
is 0 such as if there were three attributes one with a negative value
equal to the positive value and a value of 0.

If that is the case or occurs there is an update to
includes/functions/functions_lookups.php for the function
zen_has_product_attributes_values that will offer code that doesn't
suffer from the above condition.