Add functions to identify price text modification for attributes
With the addition of the ability to determine intermediate costs of a product, seemed like it was a good idea to add a little flexibility for display of starting at and/or at least text associated with product.  Have added a notifier that allows modification to identify if the text should be displayed or not.  Also added three functions, one to identify if the at least text should be displayed based on a future admin setting.  One to display the starting at text based on a future admin setting and then also a function that identifies that all methods have been chosen or are supported.  This last may be extraneous and unnecessary.

Other changes include the removal of single quotes around values that have been cast to integers within a query.
Conflicts:
	Installation files/includes/classes/dynamic_price_updater.php