Add pre-discount return price
First attempt at returning the text associated with pre-discount prices.

This is expected to support the possibility of updating one of the three possible values to be displayed on the product information page. For example if a product is on sale and the attributes are associated with (or not) the sale price, then this is supposed to offer the updated text related to the crossed out normal price and the updated final price.

Problem to still be addressed is what if a product has both a special and a sale where the product's sale price is dependent on the special price, then how does the middle "special" price get updated?