remove parameters from URL
Found that during use, if the action parameter remained on the url to support processing that when communicating with say `ajax.php` that ZC would perform `main_cart_actions` based on the full information provided.  This became an observable issue when the product had a text field identified as required, the product was submitted to the shopping cart which resulted in returning to the product information page with the action parameter still on the url and then when DPU generated the url to process, the action parameter caused ZC to attempt to process that action and an unexpected redirect to occur.

This commit prevents including the action parameter in the url for product_music_info.