Split additional text from price.
For the case where trying to update the "base" price while also updating
the final price, in the default situation the text that begins the statement
is outside the group of the price itself. To offer the most flexibility
trying to split the two into different variables/data so that can be
treated as necessary in recombination.  Work remains on this topic.