Split pretext for Discount text
The general output for ZC is to display text, then within a span the
number for the associated price when displaying a price that is to be
discounted.  This separates the prefix text so that as desired/designed
the prefix text can be separately integrated into the changed price
value that displays the pre-discount text (ie. non-sale, non-special,
but includes the base price of the options selected).