Populate whos online to show current page
Because an ajax call processes through ZC, the URI captured by ZC
reflects the ajax call which causes a loss of data associated with
who's online when viewing the current status. To address this, DPU
has had an additional operation incorporated to capture the current
page which also shows on the URI that the operation includes the ajax
call for DPU.  In this way, a store owner can see that an ajax call
was made but also shows the standard page information as populated using
zen_href_link.