Add text update for normalprice class text.
When a product is on sale or special, typically the normalprice is stricken
through though there is original text that prefixes the lined through
product price. This pushes the updated text that is before the actual.  This
allows the pre-text to be updated according to the attribute selections.

Code was added to work with javascript style modifications with a jQuery
method available but commented out.